test(ledger): consolidate advance backfill scenarios
